Feature/add tests

See merge request open-package-deploy/backend!34

+++ agent/src/main/java/org/codesystem/utility/AutoUpdateUtility.java
@@ -9,7 +9,7 @@ import java.nio.file.Files;
 public class AutoUpdateUtility {
 
     public void updateApplication() {
-        if (!Variables.PATH_UPDATE_FILE.toFile().exists() || !Variables.PATH_FILE.toFile().exists()) {
+        if (!Variables.PATH_UPDATE_FILE.toFile().exists()) {
             throw new SevereAgentErrorException("Cannot find update file");
         }
 
@@ -21,15 +21,17 @@ public class AutoUpdateUtility {
             }
         }
 
-        try {
-            Files.copy(Variables.PATH_FILE, Variables.PATH_BACKUP_FILE);
-        } catch (IOException e) {
-            throw new SevereAgentErrorException("Cannot create backup: " + e.getMessage());
-        }
-        try {
-            Files.delete(Variables.PATH_FILE);
-        } catch (IOException e) {
-            throw new SevereAgentErrorException("Cannot delete old Agent: " + e.getMessage());
+        if (Variables.PATH_FILE.toFile().exists()) {
+            try {
+                Files.copy(Variables.PATH_FILE, Variables.PATH_BACKUP_FILE);
+            } catch (IOException e) {
+                throw new SevereAgentErrorException("Cannot create backup: " + e.getMessage());
+            }
+            try {
+                Files.delete(Variables.PATH_FILE);
+            } catch (IOException e) {
+                throw new SevereAgentErrorException("Cannot delete old Agent: " + e.getMessage());
+            }
         }
 
         try {
